+++ cic-ipv-stub/src/tests/startCicCheck.test.ts
@@ -14,7 +14,7 @@ import axios from "axios";
 import { KMSClient, SignCommand } from "@aws-sdk/client-kms";
 import format from "ecdsa-sig-formatter";
 
-const testData = require("../events/startEvents.js")
+const testData = require("../events/startEvents.js");
 
 jest.setTimeout(30000);
 
@@ -44,7 +44,8 @@ process.env.REDIRECT_URI = "test.com/callback";
 process.env.JWKS_URI = "test.com/.well-known/jwks.json";
 process.env.CLIENT_ID = "test-id";
 process.env.SIGNING_KEY = "key-id";
-process.env.ADDITIONAL_KEY = "additional-key-id"
+process.env.ADDITIONAL_KEY = "additional-key-id";
+process.env.UNIQUE_ENCRYPTION_KEY = "unique-encryption-key-id";
 process.env.OIDC_API_BASE_URI = "api-target.com";
 process.env.OIDC_FRONT_BASE_URI = "test-target.com";
 
@@ -108,30 +109,30 @@ describe("Start CIC Check Endpoint", () => {
     expect(body.clientId).toBeDefined();
     expect(body.AuthorizeLocation).toBeDefined();
   });
-  
+
   describe("Sign function", () => {
     it("should sign the JWT using the correct key", async () => {
       const response = await handler(testData.startDefault);
-      const signCommandInput = kmsClient.commandCalls(SignCommand)[0].args[0].input; 
+      const signCommandInput =
+        kmsClient.commandCalls(SignCommand)[0].args[0].input;
       expect(signCommandInput.KeyId).toBe("key-id");
       expect(response.statusCode).toBe(200);
     });
 
     it("should sign a JWT using the correct key when provided with a custom payload for 'invalidKid'", async () => {
       const response = await handler(testData.startCustomInvalidSigningKey);
-      const signCommandInput = kmsClient.commandCalls(SignCommand)[0].args[0].input; 
+      const signCommandInput =
+        kmsClient.commandCalls(SignCommand)[0].args[0].input;
       expect(signCommandInput.KeyId).toBe("key-id");
       expect(response.statusCode).toBe(200);
     });
 
     it("should sign a JWT using the correct key when provided with a custom payload for 'missingKid'", async () => {
       const response = await handler(testData.startCustomMissingSigningKey);
-      const signCommandInput = kmsClient.commandCalls(SignCommand)[0].args[0].input; 
+      const signCommandInput =
+        kmsClient.commandCalls(SignCommand)[0].args[0].input;
       expect(signCommandInput.KeyId).toBe("key-id");
       expect(response.statusCode).toBe(200);
     });
-  })
+  });
 });
-
-
-
